New files added: 1. colvargrid.C,  2. colvargrid.h.  New features: 1) In the colvars components: added eigenvector.  Computes the projection of the coordinates on the NMA modes or some other vector field.  2) In metadynamics: can now change the size and resolution of the PMF grid.  Two options to do that are available:  a) by changing manually the grid's boundaries and width between one run and the other (keyword: rebinGrids), or  b) let the code itself expand the boundaries as needed (keyword: expandBoundaries).  3) In metadynamics: count the tails of those hills that lie close to the grid's edges; in some cases, this discontinuity, although right outside the PMF region, would previously result in some colvars getting stuck at the grid's edge in the long run.  4) Atom groups construction: define atoms by name that belong to different PSF segments (previously, only one PSF seg per group was allowed, otherwise atomNumbers was needed).  5) Colvars behaviour: define repulsive walls that don't coincide with the grid boundaries (make possible for the system to never leave the grid by any extent).  6) Cross time-correlation functions between two colvars (in addition to the ACFs).  7) Revised documentation: among several things, an example input file with actual keywords is shown instead of a *fake* one, and its syntax is now explained in more detail at the beginning.

+++ src/colvarparse.h
@@ -103,8 +103,8 @@ public:
                            std::vector<_type_> (0, _def_value_),        \
                          Parse_Mode const parse_mode = parse_normal)
   
-    _get_keyval_vector_proto_ (int, 0);
-    _get_keyval_vector_proto_ (size_t, 0);
+    _get_keyval_vector_proto_ (int, (int)0);
+    _get_keyval_vector_proto_ (size_t, (size_t)0);
     _get_keyval_vector_proto_ (std::string, std::string (""));
     _get_keyval_vector_proto_ (colvarmodule::real, 0.0);
     _get_keyval_vector_proto_ (colvarmodule::rvector, colvarmodule::rvector());
